+++ changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Remove method Parser.WriteUsageTo, use Parser.Usage().WriteTo
+- Add type Usage
+
 ## [0.6.1] 2021-10-17
 
 - Use os.Stdout to write parsing errors